LayoutTests/imported/w3c/web-platform-tests/content-security-policy/generic/generic-0_2_3-expected.txt

'self' fails with a different host (including sub-host e.g. foo.com as self with content from bar.foo.com)


PASS Prevents access to external scripts.
FAIL Should fire violation events for every failed violation assert_equals: expected "script-src-elem" but got "script-src"
